get_next_image is implemented inefficiently. It gets all non-displayed images into the python context and then selects an image from there. This leads to very large data structures. A better approach is to query how many unseen images are available (either all or the selected album). Then select a random number nth_image up to the nr of unseen images. Based on that number query the DB to  get the nth_image that is unseen and mark it as seen. This approach minimizes the data strutures that need to be handled in python.
It implemented random selection completely in SQL (not my suggestion with id selection). This is most performant. However, random selection is now 100% in DB realm, it could be implemented that it prefers performance  (i.e. select images in the sequence htey have been inserted into the DB) rather than a good distribuiont -- to be seen.