#1 Please fix the implementation in media_scanner.py to set the album attribute always to the lowest directory (i.e. the one containging the image) from the absolute path. This should even work for images that are in the root path of one of the media_paths.
Remove the other derived template sensors with which try to compute album names. Only one should persist: the one that shows the album attribute of the nextimage sensor. I do not need a reflection back of what I have selected, this is visible in the drop down.

The picture frame shows a broken link. Maybe it should be a URL instead of just a path? Check the documentation.

Entity not avaiable: button.next_mage

Entity not available: button.clear_image_history